# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Nuptial Plan – Wedding planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weddings:
    get:
      operationId: listWeddings
      tags: [weddings]
      summary: List all weddings
      responses:
        "200":
          description: List of weddings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wedding"
    post:
      operationId: createWedding
      tags: [weddings]
      summary: Create a new wedding file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeddingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wedding"
  /weddings/{id}:
    get:
      operationId: getWedding
      tags: [weddings]
      summary: Get a wedding by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Wedding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wedding"
        "404":
          description: Not found
    patch:
      operationId: updateWedding
      tags: [weddings]
      summary: Update a wedding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeddingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wedding"
    delete:
      operationId: deleteWedding
      tags: [weddings]
      summary: Delete a wedding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /weddings/{id}/summary:
    get:
      operationId: getWeddingSummary
      tags: [dashboard]
      summary: Get dashboard summary for a wedding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeddingSummary"
  /weddings/{id}/activity:
    get:
      operationId: getWeddingActivity
      tags: [dashboard]
      summary: Get recent activity feed for a wedding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── Vendors ──────────────────────────────────────────────────
  /weddings/{weddingId}/vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors for a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Add a vendor to a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /weddings/{weddingId}/vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor by ID
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Guests ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Wedding:
      type: object
      required: [id, names, weddingDate, venue, totalBudget, guestCount, createdAt]
      properties:
        id:
          type: number
        names:
          type: string
          description: "e.g. Sophie & James Hartwell"
        weddingDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        venue:
          type: string
        totalBudget:
          type: number
          description: Total budget in GBP pence
        guestCount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    WeddingInput:
      type: object
      required: [names, weddingDate, venue, totalBudget, guestCount]
      properties:
        names:
          type: string
        weddingDate:
          type: string
        venue:
          type: string
        totalBudget:
          type: number
        guestCount:
          type: number
        notes:
          type: string
    WeddingUpdate:
      type: object
      properties:
        names:
          type: string
        weddingDate:
          type: string
        venue:
          type: string
        totalBudget:
          type: number
        guestCount:
          type: number
        notes:
          type: string
    WeddingSummary:
      type: object
      required: [weddingId, daysUntil, confirmedGuests, totalGuests, budgetSpent, budgetTotal, tasksComplete, tasksTotal, upcomingEvents, vendorCount, pendingPaymentsTotal]
      properties:
        weddingId:
          type: number
        daysUntil:
          type: number
        confirmedGuests:
          type: number
        totalGuests:
          type: number
        budgetSpent:
          type: number
        budgetTotal:
          type: number
        tasksComplete:
          type: number
        tasksTotal:
          type: number
        upcomingEvents:
          type: number
        vendorCount:
          type: number
        pendingPaymentsTotal:
          type: number
    ActivityItem:
      type: object
      required: [id, weddingId, description, entityType, createdAt]
      properties:
        id:
          type: number
        weddingId:
          type: number
        description:
          type: string
        entityType:
          type: string
          enum: [vendor, guest, contract, payment, event, wedding]
        initials:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Vendor ───────────────────────────────────────────────
    Vendor:
      type: object
      required: [id, weddingId, name, category, status, totalAmountCents, createdAt]
      properties:
        id:
          type: number
        weddingId:
          type: number
        name:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [confirmed, awaiting_contract, deposit_paid, cancelled]
        totalAmountCents:
          type: number
        depositAmountCents:
          type: ["number", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name, category, status, totalAmountCents]
      properties:
        name:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [confirmed, awaiting_contract, deposit_paid, cancelled]
        totalAmountCents:
          type: number
        depositAmountCents:
          type: number
        contactName:
          type: string
        contactEmail:
          type: string
        notes:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [confirmed, awaiting_contract, deposit_paid, cancelled]
        totalAmountCents:
          type: number
        depositAmountCents:
          type: number
        contactName:
          type: string
        contactEmail:
          type: string
        notes:
          type: string

    # ─── Guest ────────────────────────────────────────────────
...