# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * The Nuptial Plan – Wedding planner API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemEntityType } from './activityItemEntityType';

export interface ActivityItem {
  id: number;
  weddingId: number;
  description: string;
  entityType: ActivityItemEntityType;
  /** @nullable */
  initials?: string | null;
  createdAt: string;
}